pyproject.toml
[project]
name = "approxssimate"
version = "0.1.0"
description = "Lightweight models for approximating SSIM from simpler distortion metrics"
authors = [
  {name = "Luc Trudeau"},
  {name = "Maria Martini"}
]
dependencies = [
  "numpy",
  "scipy",
  "scikit-image",
  "pillow"
]

[project.scripts]
approxssimate = "approxssimate.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
